backend changes.

+++ internal/billing/dto/request.go
@@ -15,8 +15,8 @@ type DispensedItem struct {
 	MedicineInventoryID string  `json:"medicine_inventory_id"`
 	PrescriptionItemID  string  `json:"prescription_item_id"`
 	BatchNo             string  `json:"batch_no"`
-	CurrentStockUnits   int     `json:"current_stock_units"`
-	QuantitySoldUnits   int64   `json:"quantity_sold_units"`
+	CurrentStockUnits   float64 `json:"current_stock_units"`
+	QuantitySoldUnits   float64 `json:"quantity_sold_units"`
 	UnitPriceCharged    float64 `json:"unit_price_charged"`
 	ComputedItemTotal   float64 `json:"computed_item_total"` //subtotal
 	TotalAmount         float64 `json:"total_amount"`        //totalprice
